Fix pen color change and grid sizing
Refactored the code so that the pen color is updated when the user
changes the color of the color picker.
Grid sizing was change removing separation of concerns
Done by using the Array.from to create grid of any size. and use a event
listener to update the size based on the newly selected size.
create x amount of.